+++ HoccerXO/UtilityCategories/NSMutableArray+Shuffle.m
@@ -14,7 +14,7 @@
 {
     for (NSUInteger i = 0; i < self.count - 1; ++i) {
         // Select a random element between i and end of array to swap with.
-        int n = random() % (self.count - i) + i;
+        NSUInteger n = random() % (self.count - i) + i;
         [self exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
 }
